moved functions between _quadratic_hamiltonian.py and _slater_determinants.py
